scanner-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
google-cloud-storage==2.18.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
